Ensure job skills are spaced in production
These look different in development and production. In development
they're spaced apart, but in production they're right up against one
another.

This is because in development, Slim templates are pretty-printed,
whereas in production, all whitespace is removed as a form of
minification. The whitespace is what actually gives the spacing, since
the elements are styled as inline-block by Foundation.